% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_container_dir_exists}
\alias{azure_container_dir_exists}
\title{azure_container_dir_exists}
\usage{
azure_container_dir_exists(
  container,
  dir,
  end_point = azure_store_endpoint(),
  hard_stop = FALSE,
  verbose = TRUE
)
}
\description{
azure_container_dir_exists
}
