% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_container_validation}
\alias{azure_container_validation}
\title{azure_container_validation}
\usage{
azure_container_validation(
  container,
  end_point = azure_store_endpoint(),
  check_negate = TRUE,
  hard_stop = TRUE,
  verbose = TRUE
)
}
\description{
azure_container_validation
}
