% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_file_load_rds}
\alias{azure_file_load_rds}
\title{azure_file_load_rds}
\usage{
azure_file_load_rds(
  container_path,
  container,
  end_point = azure_store_endpoint()
)
}
\description{
azure_file_load_rds
}
