% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure.R
\name{azure_file_save_rdata}
\alias{azure_file_save_rdata}
\title{azure_file_save_rdata}
\usage{
azure_file_save_rdata(
  obj,
  container,
  container_path = NULL,
  end_point = azure_store_endpoint()
)
}
\description{
azure_file_save_rdata
}
