% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{correlations}
\alias{correlations}
\title{correlations}
\usage{
correlations(
  df,
  dvs = NULL,
  ivs = NULL,
  cluster = TRUE,
  hclust_method = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty",
    "median", "centroid"),
  r_round = 2,
  p_round = 3,
  add_visualization = TRUE
)
}
\description{
correlations
}
