% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_NA_rows.R
\name{create_NA_rows}
\alias{create_NA_rows}
\title{create_NA_rows}
\usage{
create_NA_rows(df, n)
}
\arguments{
\item{df}{data.frame / tibble}

\item{n}{number of rows}
}
\description{
Returns empty rows
}
