% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driver.R
\name{driver}
\alias{driver}
\title{driver}
\usage{
driver(
  df,
  dv,
  ivs,
  subgroups = NULL,
  labels = NULL,
  engine = c("linear", "logistic"),
  shift_percentage = 0.05,
  dv_recode = c("none", "tb", "t2b", "t3b", "custom"),
  custom_car_recode_syntax = NULL
)
}
\description{
driver
}
