% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_viz_data.R
\name{dt_viz_data}
\alias{dt_viz_data}
\title{dt_viz_data}
\usage{
dt_viz_data(
  df,
  col_factor = NULL,
  col_round = NULL,
  col_round_digits = 2,
  col_percent = NULL,
  col_percent_digits = 2
)
}
\description{
dt_viz_data
}
