% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode.R
\name{field_recode}
\alias{field_recode}
\title{field_recode}
\usage{
field_recode(
  x,
  type = c("none", "tb", "t2b", "t3b", "custom"),
  custom_car_recode_syntax = NULL,
  df = NULL
)
}
\description{
field_recode
}
