% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_theme_picker.R
\name{hc_theme_picker}
\alias{hc_theme_picker}
\title{hc_theme_picker}
\usage{
hc_theme_picker(
  theme = c("538", "alone", "bloom", "chalk", "darkunica", "db", "economist",
    "elementary", "ffx", "flat", "flatdark", "ft", "ggplot2", "google", "gridlight",
    "handdrawn", "hcrt", "merge", "monokai", "null", "sandsignika", "smpl", "sparkline",
    "sparkline_vb", "superheroes", "tufte", "tufte2")
)
}
\description{
hc_theme_picker
}
