% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram.R
\name{histogram_single}
\alias{histogram_single}
\title{histogram_single}
\usage{
histogram_single(
  x,
  title = NULL,
  title_x_axis = NULL,
  title_y_axis = "Frequency",
  likert = NULL,
  arbitrary_likert_cutoff = 13,
  default_file_name = "filename",
  width = NULL,
  height = NULL,
  theme = c("538", "alone", "bloom", "chalk", "darkunica", "db", "economist",
    "elementary", "ffx", "flat", "flatdark", "ft", "ggplot2", "google", "gridlight",
    "handdrawn", "hcrt", "merge", "monokai", "null", "sandsignika", "smpl", "sparkline",
    "sparkline_vb", "superheroes", "tufte", "tufte2")
)
}
\description{
histogram_single
}
