% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_na_return.R
\name{if_na_return}
\alias{if_na_return}
\title{if_na_return}
\usage{
if_na_return(x, return_what = FALSE)
}
\arguments{
\item{x}{vector}

\item{return_what}{what to return}
}
\description{
Returns something if NA. Default to FALSe
}
