% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_missing_column.R
\name{insert_missing_column}
\alias{insert_missing_column}
\title{insert_missing_column}
\usage{
insert_missing_column(.data, ...)
}
\description{
insert_missing_column
}
\examples{
iris \%>\% insert_missing_column(Sepal.Length, mom, boo) \%>\% head()
}
