% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_init_pkgs.R
\name{install_init_pkgs}
\alias{install_init_pkgs}
\title{install_init_pkgs}
\usage{
install_init_pkgs(on_exit_restart = TRUE)
}
\arguments{
\item{on_exit_restart}{Logical. Restart R session of install.}
}
\description{
Installs the initial recommended packages
}
