% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pak.R
\name{install_pak}
\alias{install_pak}
\title{install_pak}
\usage{
install_pak(update = TRUE)
}
\arguments{
\item{update}{Logical value. If pak is installed, should it be updated if update is available?}
}
\description{
Checks for if pak is installed. If not, it installs it.
}
