% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_pkg.R
\name{install_pkg}
\alias{install_pkg}
\title{install_pkg}
\usage{
install_pkg(
  x = NULL,
  ask = FALSE,
  upgrade = FALSE,
  update_pkgs = FALSE,
  on_exit_restart = TRUE
)
}
\arguments{
\item{x}{Character vector. Vector of libraries to install}

\item{ask}{Logical. Whether to ask for confirmation when installing a different version of a package that is already installed. Installations that only add new packages never require confirmation.}

\item{upgrade}{Logical. When FALSE, the default, pak does the minimum amount of work to give you the latest version(s) of pkg. It will only upgrade dependent packages if pkg, or one of their dependencies explicitly require a higher version than what you currently have. It will also prefer a binary package over to source package, even it the binary package is older.}

\item{update_pkgs}{Logical. Update all current packages.}

\item{on_exit_restart}{Logical. Restart R session of install.}
}
\description{
Install R libraries
}
