% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installed_pkgs.R
\name{installed_pkgs}
\alias{installed_pkgs}
\title{installed_pkgs}
\usage{
installed_pkgs(pkg)
}
\arguments{
\item{pkg}{Character vector of packages to check to see if installed}
}
\description{
Checks if package is installed.
}
