% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_truthy.R, R/logicals.R
\name{is_truthy}
\alias{is_truthy}
\title{is_truthy}
\usage{
is_truthy(x)

is_truthy(x)
}
\arguments{
\item{x}{An expression whose truthiness value we want to determine}
}
\description{
is_truthy

Mirror of shiny::isTruthy
}
