% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left.R
\name{left}
\alias{left}
\title{left}
\usage{
left(x, n = 1)
}
\arguments{
\item{x}{Character / string of single length}

\item{n}{Integer of how many n characters to return}
}
\description{
Returns left n characters of single string
}
