% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legoland.R
\name{legoland}
\alias{legoland}
\title{legoland}
\usage{
legoland(x = c("all", "sean", "alina", "tyler", "guestservices", "services"))
}
\arguments{
\item{x}{character}
}
\description{
Returns legoland passholder codes
}
