% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litify_helpers.R
\name{lit_create_query}
\alias{lit_create_query}
\title{lit_create_query}
\usage{
lit_create_query(
  from = work::litify_dict[["object"]],
  select = NULL,
  where = NULL,
  limit = NULL,
  check_select = TRUE,
  check_from = TRUE
)
}
\arguments{
\item{from}{ojbects in litify that are serviced in work::litify_dict}

\item{select}{fields in the litify object. If NULL, all fields selected}

\item{where}{list with where conditions}

\item{limit}{numeric that limits the rows returned in the query}
}
\description{
Returns a soql querry
}
