% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_case_type.R
\name{lit_get_case_type}
\alias{lit_get_case_type}
\title{lit_get_matter_team}
\usage{
lit_get_case_type(
  cases,
  cases_field = c("id_case_type", "id_intake", "id_matter", "case", "case_type"),
  limit = NULL
)
}
\arguments{
\item{cases}{vector of case numbers}

\item{cases_field}{field type the cases are referring to}

\item{limit}{Limit for the query}
}
\description{
queries litify for matter team data
}
