% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litify_helpers.R
\name{lit_get_clip}
\alias{lit_get_clip}
\title{lit_get_clip}
\usage{
lit_get_clip(write = TRUE, write_path = "~/Downloads/")
}
\arguments{
\item{write}{logical}

\item{write_path}{character to write folder path}
}
\description{
Gets clipboard clip from litify soql all field extension
}
