% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litify_helpers.R
\name{lit_get_dictionary_from}
\alias{lit_get_dictionary_from}
\title{lit_get_dictionary_from}
\usage{
lit_get_dictionary_from(
  from = work::litify_dict[["object"]],
  what = c("dictionary", "relationships")
)
}
\arguments{
\item{from}{ojbects in litify that are serviced in work::litify_dict}

\item{what}{fields srom the object, relationships, or both}
}
\description{
Returns dictionaries from litify objects
}
