% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_drops.R
\name{lit_get_drops}
\alias{lit_get_drops}
\title{lit_get_drops}
\usage{
lit_get_drops(
  start_date = as.Date("2021-01-01"),
  parallel_process = TRUE,
  add_lead_gen_source = FALSE
)
}
\description{
queries litify for drop and subout data
}
