% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litify_helpers.R
\name{lit_get_fields_from}
\alias{lit_get_fields_from}
\title{lit_get_fields_from}
\usage{
lit_get_fields_from(from = work::litify_dict[["object"]])
}
\arguments{
\item{from}{ojbects in litify that are serviced in work::litify_dict}
}
\description{
Returns fields from litify objects
}
