% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_intake.R
\name{lit_get_intake}
\alias{lit_get_intake}
\title{lit_get_intake}
\usage{
lit_get_intake(
  cases = NULL,
  limit = NULL,
  cases_field = c("case", "id_intake", "id_matter", NULL),
  question_filter = c("Type of Accident.", "Type of Accident")
)
}
\description{
queries litify for intake data
}
