% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_intake_questionnaire.R
\name{lit_get_intake_questionnaire}
\alias{lit_get_intake_questionnaire}
\title{lit_get_intake_questionnaire}
\usage{
lit_get_intake_questionnaire(
  cases = NULL,
  limit = NULL,
  cases_field = c(NULL, "case", "id_intake", "id_matter"),
  question_filter = NULL,
  clean_cols = TRUE,
  chunks = 700
)
}
\description{
queries litify for intake questionnaire
}
