% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_matter.R
\name{lit_get_matter_DEPRECATE}
\alias{lit_get_matter_DEPRECATE}
\title{lit_get_matter_DEPRECATE}
\usage{
lit_get_matter_DEPRECATE(
  cases = NULL,
  limit = NULL,
  cases_field = c("case", "id_intake", "id_matter", NULL)
)
}
\description{
queries litify for matter data
}
