% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_matter_history.R
\name{lit_get_matter_history}
\alias{lit_get_matter_history}
\title{lit_get_matter_history}
\usage{
lit_get_matter_history(id_matter = NULL, parallel_process = TRUE)
}
\arguments{
\item{id_matter}{vector of matter ids}
}
\description{
queries litify for matter history
}
