% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_matter_team.R
\name{lit_get_matter_team}
\alias{lit_get_matter_team}
\title{lit_get_matter_team}
\usage{
lit_get_matter_team(
  cases = NULL,
  cases_field = "Needles_CaseID__c",
  add_links = FALSE,
  parallel_process = TRUE
)
}
\arguments{
\item{cases}{vector of case numbers}
}
\description{
queries litify for matter team data
}
