% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_get_resolutions.R
\name{lit_get_resolutions}
\alias{lit_get_resolutions}
\title{lit_get_resolutions}
\usage{
lit_get_resolutions(
  cases = NULL,
  limit = NULL,
  apply_translate_limits = TRUE,
  add_links = list(link_resolution = "id_resolution", link_matter = "id_matter"),
  nested_structure = list(by = "id_matter", data_name = "resolution_data", id_matter =
    "resolution_data \%>\% purrr::map_vec(function(x)x[1,\\"id_matter\\"]) \%>\% unlist()",
    case =
    "resolution_data \%>\% purrr::map_vec(function(x)x[1,\\"case\\"]) \%>\% unlist()",
    resolution_count = "resolution_data \%>\% purrr::map_vec(nrow)")
)
}
\arguments{
\item{cases}{vector of case numbers}
}
\description{
queries litify for resolution data
}
