% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lit_process_date_filed.R
\name{lit_process_date_filed}
\alias{lit_process_date_filed}
\title{lit_process_date_filed}
\usage{
lit_process_date_filed(x, remove_input_dates = TRUE, insert_if_missing = FALSE)
}
\description{
process date filed
}
