% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litify_helpers.R
\name{lit_put_dictionary_in_pkg}
\alias{lit_put_dictionary_in_pkg}
\title{lit_put_dictionary_in_pkg}
\usage{
lit_put_dictionary_in_pkg(where = "~/Downloads/Map.xlsx", overwrite = TRUE)
}
\arguments{
\item{where}{character to write workbook path. Must be xlsx file.}
}
\description{
Writes litify dictionary to work package.
}
