% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/litify_helpers.R
\name{lit_run_query}
\alias{lit_run_query}
\title{lit_run_query}
\usage{
lit_run_query(
  from = work::litify_dict[["object"]],
  select = NULL,
  where = NULL,
  limit = NULL,
  check_select = TRUE,
  check_from = TRUE,
  query = NULL,
  return_dictionary = FALSE
)
}
\arguments{
\item{from}{ojbects in litify that are serviced in work::litify_dict}

\item{select}{fields in the litify object. If NULL, all fields selected}

\item{where}{list with where conditions}

\item{limit}{numeric that limits the rows returned in the query}

\item{query}{noql query string}

\item{return_dictionary}{logical.  Will not return if you use the query parameter}
}
\description{
Returns query from litify objects
}
