% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names_clean.R
\name{names_clean}
\alias{names_clean}
\title{names_clean}
\usage{
names_clean(x, make_lowercase = TRUE)
}
\arguments{
\item{x}{named df or character vector}
}
\description{
Returns dictionaries from litify objects
}
