% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_col_names.R
\name{other_col_names}
\alias{other_col_names}
\title{other_col_names}
\usage{
other_col_names(.data, ...)
}
\description{
other_col_names
}
\examples{
iris \%>\% other_col_names(Species, hi, Petal.Length)
}
