% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_case_costs.R
\name{process_case_costs}
\alias{process_case_costs}
\title{process_case_costs}
\usage{
process_case_costs(where, sheet_name = "Sheet1")
}
\arguments{
\item{where}{Location of the QB export}
}
\description{
Process case costs
}
