% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_scrape.R
\name{process_scrape}
\alias{process_scrape}
\title{process_scrape}
\usage{
process_scrape(
  type = c("glassdoor"),
  where = NULL,
  subtitle = "wilshire-law-firm",
  ...
)
}
\arguments{
\item{type}{vector of glassdoor company name.}

\item{company_number}{vector of glassdoor company number.}

\item{gd_base}{vector of glassdoor address base. Default to "https://www.glassdoor.com/Reviews/".}

\item{url_instructions}{vector of glassdoor url instructions. Default to "sort.sortType=RD&sort.ascending=false&filter.iso3Language=eng"}
}
\description{
Returns logical if all values in vector are unique
}
