% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pkgs.R
\name{remove_pkgs}
\alias{remove_pkgs}
\title{remove_pkgs}
\usage{
remove_pkgs(x = NULL)
}
\arguments{
\item{x}{Character vector. If x is NULL, remove all non-essential packages.}
}
\description{
Remvoes packages
}
