% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_col.R
\name{rename_col}
\alias{rename_col}
\title{rename_col}
\usage{
rename_col(.data, ..., .select = FALSE, .distinct = FALSE)
}
\description{
rename_col
}
\examples{
iris \%>\% rename_col(
s_len = Sepal.Length,
tom = mom,
p_len = Petal.Length,
foo = boo
) \%>\% head()
}
