% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restart.R
\name{restart}
\alias{restart}
\title{restart}
\usage{
restart(keep = FALSE, clean = TRUE, restart_session = TRUE)
}
\arguments{
\item{keep}{Logical or character vector. TRUE is keep all objects on restart. FALSE is drop all objects on restart. Character vector specifies which objects to keep.}

\item{clean}{Logical. If TRUE, gc() is executed after restart. Note that the gc() is executed before the session is restarted.}

\item{restart_session}{Logical. IF true, the R session will be restarted}
}
\description{
Custom restart function that exectues common tasks
}
