% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/right.R
\name{right}
\alias{right}
\title{right}
\usage{
right(x, n = 1)
}
\arguments{
\item{x}{Character / string of single length}

\item{n}{Integer of how many n characters to return}
}
\description{
Returns right n characters of single string
}
