% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_r_environment.R
\name{set_r_environment}
\alias{set_r_environment}
\title{set_r_environment}
\usage{
set_r_environment(
  type = c("custom", "git_local_dir"),
  x = NULL,
  overwrite = FALSE,
  on_exit_restart = TRUE
)
}
\arguments{
\item{type}{Type of set to do}

\item{x}{Character vector. Lines to add to the R environment file.}
}
\description{
Set the R environment
}
