% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_r_profile.R
\name{set_r_profile}
\alias{set_r_profile}
\title{set_r_profile}
\usage{
set_r_profile(
  type = c("init", "my", "custom"),
  x = NULL,
  overwrite = TRUE,
  on_exit_restart = TRUE
)
}
\description{
setup the local r profile file
}
