% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_scrub.R
\name{str_scrub}
\alias{str_scrub}
\title{str_scrub}
\usage{
str_scrub(
  x,
  replacement = "_",
  make_lowercase = TRUE,
  return_on_error = TRUE,
  fill_na = NULL,
  fill_null = NULL,
  keep = NULL,
  remove_all_spaces = TRUE,
  remove_utf8 = TRUE
)
}
\arguments{
\item{x}{Character / string vector}

\item{replacement}{Integer of how many n characters to return}
}
\description{
Returns a scrubbed string
}
