% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_limits.R
\name{translate_limits}
\alias{translate_limits}
\title{translate_limits}
\usage{
translate_limits(x)
}
\arguments{
\item{x}{character vector to translate}
}
\description{
returns scrubbed limit vector
}
