% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_names.R
\name{translate_names}
\alias{translate_names}
\title{translate_names}
\usage{
translate_names(x, type = c("cm", "attorney", "negotiator"))
}
\arguments{
\item{x}{character vector to translate}
}
\description{
returns scrubbed names vector
}
