% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_cases_equal.R
\name{where_cases_equal}
\alias{where_cases_equal}
\title{where_cases_equal}
\usage{
where_cases_equal(x, api_field = "Needles_CaseID__c")
}
\arguments{
\item{x}{value vector}

\item{api_field}{name of api field}
}
\description{
returns sql / nosql syntax for case logic, typically used in where command.
}
